@@ -14,27 +14,27 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^14.2.0",
-    "@angular/cdk": "^14.2.2",
-    "@angular/common": "^14.2.0",
-    "@angular/compiler": "^14.2.0",
-    "@angular/core": "^14.2.0",
-    "@angular/forms": "^14.2.0",
-    "@angular/material": "^14.2.2",
-    "@angular/platform-browser": "^14.2.0",
-    "@angular/platform-browser-dynamic": "^14.2.0",
-    "@angular/router": "^14.2.0",
+    "@angular/animations": "^16.0.5",
+    "@angular/cdk": "^16.0.4",
+    "@angular/common": "^16.0.5",
+    "@angular/compiler": "^16.0.5",
+    "@angular/core": "^16.0.5",
+    "@angular/forms": "^16.0.5",
+    "@angular/material": "^16.0.4",
+    "@angular/platform-browser": "^16.0.5",
+    "@angular/platform-browser-dynamic": "^16.0.5",
+    "@angular/router": "^16.0.5",
     "apollo-angular": "^4.2.1",
     "graphql": "^16.6.0",
     "rxjs": "~7.5.0",
     "tslib": "^2.3.0",
-    "zone.js": "~0.11.8"
+    "zone.js": "~0.13.0"
   },
   "devDependencies": {
-    "@angular-builders/jest": "^14.1.0",
-    "@angular-devkit/build-angular": "^14.2.3",
-    "@angular/cli": "~14.2.3",
-    "@angular/compiler-cli": "^14.2.0",
+    "@angular-builders/jest": "^16.0.0",
+    "@angular-devkit/build-angular": "^16.0.5",
+    "@angular/cli": "^16.0.5",
+    "@angular/compiler-cli": "^16.0.5",
     "@azure/static-web-apps-cli": "^1.0.6",
     "@ngx-env/builder": "^16.0.0",
     "@types/jest": "^29.2.4",
@@ -43,7 +43,7 @@
     "@typescript-eslint/parser": "^5.48.1",
     "eslint": "^8.31.0",
     "eslint-config-prettier": "^8.6.0",
-    "jest": "^28.1.3",
+    "jest": "^29.5.0",
     "typescript": "^4.8.4"
   }
 }
