@@ -1,4 +1,4 @@
-import { Component, OnInit } from "@angular/core";
+import { Component, OnInit, inject } from "@angular/core";
 import { CommonModule } from "@angular/common";
 import { MatButtonModule } from "@angular/material/button";
 import { MatDividerModule } from "@angular/material/divider";
@@ -22,11 +22,9 @@ export class HomepageComponent implements OnInit {
   blogUrl = environment.blogUrl;
   noresults = "Searching for featured listings. Please wait...";
 
-  constructor(
-    private listingService: ListingService,
-    private favoriteService: FavoriteService,
-    private userService: UserService,
-  ) {}
+  private listingService = inject(ListingService);
+  private favoriteService = inject(FavoriteService);
+  private userService = inject(UserService);
 
   async ngOnInit() {
     this.user = await this.userService.currentUser();
