@@ -1,4 +1,4 @@
-import { Injectable } from "@angular/core";
+import { Injectable, inject } from "@angular/core";
 import { Resolve } from "@angular/router";
 import { BehaviorSubject } from "rxjs";
 import { LocalStorageService } from "../local-storage/local-storage.service";
@@ -16,7 +16,7 @@ export class UserService implements Resolve<User> {
   private readonly userSource = new BehaviorSubject<User>(this.guestUser());
   readonly user$ = this.userSource.asObservable();
 
-  constructor(private localStorageService: LocalStorageService) {}
+  private localStorageService = inject(LocalStorageService);
 
   async resolve(): Promise<User> {
     return await this.currentUser();
